refactor: 등록하기 버튼과 업로드 input을 분리
